+++ src/britive/britive.py
@@ -35,7 +35,7 @@ from .policies import Policies
 from .secrets_manager import SecretsManager
 from .notification_mediums import NotificationMediums
 from .workload import Workload
-
+from .system.system import System
 
 BRITIVE_TENANT_ENV_NAME = 'BRITIVE_TENANT'
 BRITIVE_TOKEN_ENV_NAME = 'BRITIVE_API_TOKEN'
@@ -182,6 +182,7 @@ class Britive:
         self.secrets_manager = SecretsManager(self)
         self.notification_mediums = NotificationMediums(self)
         self.workload = Workload(self)
+        self.system = System(self)
 
     @staticmethod
     def source_federation_token_from(provider: str, tenant: str = None, duration_seconds: int = 900) -> str: